Support for both evfevent remote and local
